+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
+- Added: Add support to the attribute `hidden` on flyout block elements to hide content for when JavaScript and CSS is unavailable.
 - Added: Support to target flyout blocks using the `href` of a link attribute.
 - Added: Set trigger elements `role` to `button`.
 - Added: Support to release `disabled` and `aria-hidden` on trigger elements at initialization.